mostly finished components/ReactTestStatements to ts
